Tidy acceptance tests for consistency
- Remove unused steps
- Remove unused selectors
- Remove unused commands
