+++ test/acceptance/features/step_definitions/details.js
@@ -80,24 +80,6 @@ const assertTableContent = async function (tableSelector, expectedData, tableTyp
   }
 }
 
-Then(/^details heading should contain what I entered for "(.+)" field$/, async function (fieldName) {
-  await Details
-    .waitForElementPresent('@heading')
-    .assert.containsText('@heading', this.state[fieldName])
-})
-
-Then(/^details heading should contain "(.+)"$/, async (value) => {
-  await Details
-    .waitForElementPresent('@heading')
-    .assert.containsText('@heading', value)
-})
-
-Then(/^details content heading should contain "(.+)"$/, async (value) => {
-  await Details
-    .waitForElementPresent('@contentHeading')
-    .assert.containsText('@contentHeading', value)
-})
-
 Then(/^details view data for "(.+)" should contain what I entered for "(.+)" field$/, async function (detailsItemName, fieldName) {
   const detail = await Details.getDetailFor(detailsItemName)
 
@@ -116,28 +98,6 @@ Then(/^details view data for "(.+)" should contain "(.+)"$/, async (detailsItemN
     .waitForElementPresent(detail.selector)
     .assert.containsText(detail.selector, value)
     .useCss()
-})
-
-Then(/^there should be a local nav$/, async (dataTable) => {
-  const expectedLocalNav = dataTable.hashes()
-
-  await Details.api.elements('css selector', '.c-local-nav__link', (result) => {
-    client.expect(result.value.length).to.equal(expectedLocalNav.length)
-  })
-
-  for (const row of expectedLocalNav) {
-    const localNavItemSelector = Details.getLocalNavItemSelector(row.text)
-    await Details
-      .api.useXpath()
-      .waitForElementPresent(localNavItemSelector.selector)
-      .assert.visible(localNavItemSelector.selector)
-      .useCss()
-  }
-})
-
-Then(/^there should not be a local nav$/, async () => {
-  await Details
-    .assert.elementNotPresent('@localNav')
 })
 
 Then(/^the (.+) key value details are displayed$/, async function (tableTitle, dataTable) {
